CC-8: modified product database.

+++ product/models.py
@@ -3,7 +3,7 @@ from django.db import models
 
 class Product(models.Model):
     name = models.CharField(max_length=50)
-    calories = models.IntegerField()
+    calories = models.FloatField()
 
     def __str__(self):
         return f"100 g/ml of {self.name} = {self.calories}kcal"
